chore(spec): use jest-environment-hops instead of local environment
