+++ packages/spec/integration/development-proxy/package.json
@@ -14,7 +14,7 @@
     ]
   },
   "jest": {
-    "testEnvironment": "../../helpers/env.js",
+    "testEnvironment": "jest-environment-hops",
     "setupFilesAfterEnv": [
       "../../jest.setup.js"
     ]
