+++ packages/spec/integration/graphql-mock-server/__tests__/develop.js
@@ -1,3 +1,5 @@
+const { handleConsoleOutput } = require('../../../helpers');
+
 describe('graphql mock server', () => {
   let url;
 
@@ -8,6 +10,7 @@ describe('graphql mock server', () => {
 
   it('renders a mocked quote', async () => {
     const { page, getInnerText } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     expect(await getInnerText('p')).toBe('Hello World');
@@ -17,6 +20,7 @@ describe('graphql mock server', () => {
 
   it('renders an email through a custom resolver', async () => {
     const { page, getInnerText } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     expect(await getInnerText('cite')).toBe('email@example.com');
