+++ packages/spec/integration/graphql-no-ssr/__tests__/develop.js
@@ -1,3 +1,5 @@
+const { handleConsoleOutput } = require('../../../helpers');
+
 describe('graphql mock server without SSR', () => {
   let url;
 
@@ -8,6 +10,7 @@ describe('graphql mock server without SSR', () => {
 
   it('requests data on the client & not on the server', async () => {
     const { page, getElementByText } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     const response = await page.goto(url, { waitUntil: 'networkidle2' });
     const html = await response.text();
     const loading = '<strong id="loading">loading...</strong>';
