+++ packages/spec/integration/graphql-no-ssr/package.json
@@ -13,7 +13,7 @@
     "graphqlMockSchemaFile": "<rootDir>/mock-schema.js"
   },
   "jest": {
-    "testEnvironment": "../../helpers/env.js",
+    "testEnvironment": "jest-environment-hops",
     "setupFilesAfterEnv": [
       "../../jest.setup.js"
     ]
