+++ packages/spec/integration/graphql/package.json
@@ -13,7 +13,7 @@
     "graphqlUri": "https://some.graphql/api"
   },
   "jest": {
-    "testEnvironment": "../../helpers/env.js",
+    "testEnvironment": "jest-environment-hops",
     "setupFilesAfterEnv": [
       "../../jest.setup.js"
     ]
