+++ packages/spec/integration/hot-module-reload/package.json
@@ -6,7 +6,7 @@
     "node": "12 || 14 || 15"
   },
   "jest": {
-    "testEnvironment": "../../helpers/env.js",
+    "testEnvironment": "jest-environment-hops",
     "setupFilesAfterEnv": [
       "../../jest.setup.js"
     ]
