+++ packages/spec/integration/postcss/__tests__/develop.js
@@ -1,3 +1,5 @@
+const { handleConsoleOutput } = require('../../../helpers');
+
 describe('react-postcss', () => {
   let url;
 
@@ -8,6 +10,7 @@ describe('react-postcss', () => {
 
   it('styles when served in development mode', async () => {
     const { page } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     const { position, fontFamily } = await page.evaluate(() => {
@@ -28,6 +31,7 @@ describe('react-postcss', () => {
 
   it('does not add css file to template', async () => {
     const { page } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.setJavaScriptEnabled(false);
 
     await page.goto(url, { waitUntil: 'networkidle2' });
@@ -40,6 +44,7 @@ describe('react-postcss', () => {
 
   it('supports global un-hashed CSS classnames', async () => {
     const { page } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     const { backgroundColor, animationDuration } = await page.evaluate(() => {
