+++ packages/spec/integration/pwa/package.json
@@ -13,7 +13,7 @@
     ]
   },
   "jest": {
-    "testEnvironment": "../../helpers/env.js",
+    "testEnvironment": "jest-environment-hops",
     "setupFilesAfterEnv": [
       "../../jest.setup.js"
     ]
