+++ packages/spec/integration/redux/__tests__/develop.js
@@ -1,4 +1,5 @@
 const urlJoin = require('url-join');
+const { handleConsoleOutput } = require('../../../helpers');
 
 describe('redux development server', () => {
   let url;
@@ -10,6 +11,7 @@ describe('redux development server', () => {
 
   it('has default state', async () => {
     const { page, getInnerText } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     const count = await getInnerText('counter');
@@ -21,6 +23,7 @@ describe('redux development server', () => {
 
   it('increments the counter on click', async () => {
     const { page, getInnerText } = await createPage();
+    page.on('console', (msg) => handleConsoleOutput(msg));
     await page.goto(url, { waitUntil: 'networkidle2' });
 
     await page.click('button');
@@ -36,6 +39,7 @@ describe('redux development server', () => {
   describe('route matching action creators', () => {
     it('will be executed if the route matches', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/increment'), {
         waitUntil: 'networkidle2',
       });
@@ -49,6 +53,7 @@ describe('redux development server', () => {
 
     it('will not be executed if the location hash changes', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       const hashValue = '#some-hash';
       await page.goto(urlJoin(url, '/increment'), {
         waitUntil: 'networkidle2',
@@ -66,6 +71,7 @@ describe('redux development server', () => {
 
     it('will be executed if a search query or hash is present', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/increment?foo=bar#hash'), {
         waitUntil: 'networkidle2',
       });
@@ -79,6 +85,7 @@ describe('redux development server', () => {
 
     it('will be awaited if they are promises', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/increment-fetch'), {
         waitUntil: 'networkidle2',
       });
@@ -92,6 +99,7 @@ describe('redux development server', () => {
 
     it('will be provided with expected "param" string', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/param-test/foo'), {
         waitUntil: 'networkidle2',
       });
@@ -105,6 +113,7 @@ describe('redux development server', () => {
 
     it('will be provided with the expected "location" object', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/location-test?foo=bar'), {
         waitUntil: 'networkidle2',
       });
@@ -118,6 +127,7 @@ describe('redux development server', () => {
 
     it('will be provided with the expected "match" object', async () => {
       const { page, getInnerText } = await createPage();
+      page.on('console', (msg) => handleConsoleOutput(msg));
       await page.goto(urlJoin(url, '/match-test/foobar'), {
         waitUntil: 'networkidle2',
       });
